feat(account/security): refactor session fetching into a custom hook for improved code organization
